Greet-Tweet-Know 1.0.0a9: status bar & display tweaks
I'm trying to get Greet-Tweet-Know to support special actions on for tweets, ie [Ctrl+Return] starting a repley to the selected tweet.  But handling all the need keyboard commands, like up & down, are taking longer than I'd hoped but I'm still working on that.  Until I have those working the menu shortcuts still work.

I'm displaying this shortcuts in the status bar when a tweet is selected, unless a different message is set.  I'm working at making these new messages only temporary, eg they'll display for 10 seconds & return to the shortcuts message, but I've still not found a way to do that, reliably that doesn't impact performance.  So for now a different tweet has to be selected before the shortcut message is redisplayed.

Committed on: 2009-03-27 @ 11:07:52 PM