Greet-Tweet-Know 0.0.eCa9: Friends functions finished, extended tweet buttons added, and support for Twitter favorites added.
I know I said it last time...but this time I mean it, this is Greet-Tweet-Know's 1st official release.  Packaging help would be warmly welcomed.

We finished our menu entries for manually following, unfollowing, and blocking users.  We have done a lot of continued work on revamping Greet-Tweet-Know's interface.  We're still just using GTK so it can only be so pretty - which means pretty ugly.  So we're trying to extend its UI's functionality as much as we can before we begin to really work on recreating Greet-Tweet-Know's full interface.  We'll be doing this with a graphic's library that supports animations, transparancies, and much more.  We'll be using a game library - but that's for what's to come.  Which is so much more than what it is...or was ever intended to be.  LOL as is we've already re-wrote, re-factored, and re-designed at least 700f the code we orginally in herited.  At least its slowly look less ugly & working much better.

I added buttons to Greet-Tweet-Know's extended Tweet view.  These buttons are for replying, retweeting, creating a DM, and for marking the Tweet as a favorite.

I have put off work on my 'combined viem' for a while. There are a lot of lil tweaks I want to do 1st.  I have added Twitter's Favorites timeline to Greet-Tweet-Know's 'View Tweets' menu.  But as for my 'combined' timeline I've paused workins on it until I extend the tweet view so it has an improved list view.  I'll be building it from a gtkbuild file or using another graphics library to refactoring GtK's tweet-list/timeline viewer.

I'm not @version-1 yet but every thing in Greet-Tweet-Know is functional & stable.  Even though several of the bug fixes have been kinda embarrassing I'm glad they're fixed & I'm glad to have added every thing that I have.

Though I have not yet reached version one yet I'm not focusing on Greet-Tweet-Know.  I really want & need to have time to focus on my game & my podcast. Spending 12+hrs each day on Greet-Tweet-Know isn't giving me enough time to do all the different things that my interest pull me in.  Not to mention I hardly ever get to Tweet when I'm spending all day working on my Twitter client.

But we'll do what makes us happy & lots lately that's doing stuff for Greet-Tweet-Know.  But that's about to change and we're gonna be doing a lot more stuff.

Committed on: 2009-04-12 @ 02:25:36 AM