#00BF01a19
This is my 1st release with my new best friends feature in place.  Don't worry there are a ton of improvements are still on the way.  But now every thing's in place & working.

This is my most important feature I feel I've ever programmed in any application.  This is so I'll be able to keep up with my few friends lives, as best as I can, and still work on my projects.

This wil be especially important as I work on my games more & more. But for my next several weeks I'll be working on adding a lot of *uber-shiny* stuff to *top off* my best friends feature.

Some of the shiny I want to add is stuff like making news tabs loading in the 'background' or 'foreground' as a preference.  Right now new tabs open in the 'background', like Firefox default action.

First though I'll be adding a check to notify you of all of your best friends new updates since your last login.  I'm also going to be adding a 'new updates' column the 'best friends' tree view. I'll also be making it so best friends with unread update get their name's marked-up with a bold tag so its easy to see who's updates you've missed.

I have a ton of shiny lil improvements I'll be finishing up to truly feel like get2gnow's 'best friend' feature is as good as I want it to be.

But for now - here it is.  And its totally shiny :-)

Oh & I've add a ton of other UI improvements and new features including a new view menu. It drastically improves how one's able to customize get2gnow's appearance allowing it to be viewed in a much more compact view.  More shiny stuff.

Lastly there are *tons* of bug fixes & patches.  I think I've even resolved the random libsoup issues that was causing segfaults.  I've done a ton of stuff & still have a ton more I want to do.

Committed on : 2009-09-21 @ 05:03:38 AM