#010001a01
*uber-w00t!* get2gnow's nightmarish account management bugs are no more, they are dead, toast, fried.  At least all that I know of.  Yeas I'm abs sure you could break something if you really really tried(i.e. my previous statement wasn't a dare) but if you're going to take it as one than please at least let me know what you've done.

Otherwise for future proofing, eextensiblity, better thread support, & to just create better code I will still be refactoring my accounts / online services system.  This is something I've already been working for about one week or so.  I will continue to work re-factoring my 'online-service' as well as refactoring my User & UserStatus objects.  However at least at this moment I can breath a sigh of relief because this work is not, after all, required for get2gnow to be usable.  It simply required me to have a night's sleep & a clear look at the problem.

Now on to my re-factoring which will include, as an end result, support for Twitter searches and, possibly, StatusNet groups as well.

On to my day w/a clear head & intentions of remaining calm - even in the face of seemingly unsolvable bugs. LOL yeah right, as if. I say that & I'd like to say I'd even like to follow throughh with it.  However the only way I could possibly not get emotional when something I'm creating isn't working 'just right' is iif I weren't passionate about my work.  And, as is the case with all of my art works, I am passionate about what I create , how beautiful it is - which extends to how beautifully my programs work, when that is my art form & form of expression.

So yes, LOL, in my future I'll continue to be a moody, ranty, bitchy open source artist.  But when all is said and done - when you're running & experiencing get2gnow - you'll also agree that it truly is a *piece of art*.

Committed on : 2009-10-07 @ 01:42:54 PM