#010001a08
I've implement a fix for getting get2gnow to support libsoup > 2.16.1.  It turned out to be an easier to 'work around' fix than I'd even imagined.  I simply had to implement a 'loading' time before timelines are loaded.  This gives newer versions of libsoup a chance to 'settle' and for one's connection to complete before get2gnow sends multiple requests to libsoup.

Committed on : 2009-10-08 @ 09:51:47 PM