#010001a26
I've implement major improvements to get2gnow's UpdateViewer so it now uses its tree_model_sort and list_store in combination so whenever it is updating the list store's content it only has to make one pass over the tree_model_sort and uses a new column to determine which indices to use when removing or updating rows in the list_store.

I've also made impovement to how the control-panel & update area.  There where certain distros, or possibly GTK+ or libsexy versions, which were not displaying & enabling get2gnow's update entry 'text box', i.e. SexyTextEntry.  I've fixed this - though why it was a bug is not something I could replicate but at least I've been able to figure what was causing it.

Committed on : 2009-10-17 @ 10:42:34 AM