#010001a40
I've improved get2gnow's queue so requests are re-processed every millisecond instead of once each second.  When a lot of requests are queued this will increase the resources that are used but it drastically improves how quickly network requests are processed.  So a slight increase in get2gnow's foot print, in rare circumstances.  Its a worth while trade off for drastic improvements in get2gnow's performance over all.  Now the fact that requests are queued shouldn't even be noticed.

I also finished refactoring how & when get2gnow calculates how often updates are reloaded so its not recalculating reload values repeatadly.  It only calculates minutes for reloading when needed.  This slight improvement makes it so get2gnow only calculates a timelines reloading time once, instead of three times like it was.

Lastly I've completely refactored my OnlineServiceWrapper object, which is used for processing & queing network requests.  Every thing is centeralized and ready to be turned into a gobject, if I chose to but, for artistic reasons, I've decided, for now, that I won't be implementing gobject in as much source as I was previously planning on doing.

There's not a ton in this commit - but what is here has uber huge pay offs.

Committed on : 2009-10-23 @ 03:35:04 AM